requirements.txt
anyio==4.3.0
certifi==2024.2.2
cleo==2.1.0
crashtest==0.4.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
notion-client==2.2.1
python-dotenv==1.0.1
rapidfuzz==3.6.2
sniffio==1.3.1
typing_extensions==4.10.0
